requirements.txt
fastapi
uvicorn
starlette
jinja2
python-dotenv
mysql-connector-python
bcrypt
opencv-python
numpy
tensorflow
keras
mediapipe
face-recognition
insightface
python-multipart
itsdangerous
dnspython
torch
torchvision
pillow
pyyaml
onnxruntime-gpu